OCRService/requirements.txt
# PaddleOCR 服务依赖
paddleocr>=2.7.0
paddlepaddle>=2.5.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0